# laravel-with-jwt/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "operationId": "be2e78e5f93182d4403f839cc516d312",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success, return new token"
                    },
                    "401": {
                        "description": "Unauthorized, return error message"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User registration",
                "operationId": "58fe5a217ef88ae75e9b530a5e5817d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "minLength": 6
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm Password",
                                        "type": "string",
                                        "minLength": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "401": {
                        "description": "Error, return error message"
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "operationId": "1a1e54a850ca4218b0388d92093ab652",
                "responses": {
                    "200": {
                        "description": "Success, return new token"
                    },
                    "401": {
                        "description": "Unauthorized, return error message"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "operationId": "0b366af9c6842623aab9c8827c519b89",
                "responses": {
                    "200": {
                        "description": "Success, return logout message"
                    },
                    "401": {
                        "description": "Unauthorized, return error message"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update user password",
                "operationId": "e3b02892439fc1f051d8ecef3ba9055e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "current_password": {
                                        "description": "Current password",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "New password",
                                        "type": "string",
                                        "minLength": 6
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm new password",
                                        "type": "string",
                                        "minLength": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized, return error message"
                    },
                    "500": {
                        "description": "Unable to update password, return error message"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}